+++ packages/workers-shared/router-worker/src/index.ts
@@ -2,10 +2,11 @@ import { PerformanceTimer } from "../../utils/performance";
 import { setupSentry } from "../../utils/sentry";
 import { mockJaegerBinding } from "../../utils/tracing";
 import { Analytics, DISPATCH_TYPE } from "./analytics";
+import { applyConfigurationDefaults } from "./configuration";
 import type AssetWorker from "../../asset-worker/src/index";
 import type {
 	JaegerTracing,
-	RoutingConfig,
+	RouterConfig,
 	UnsafePerformanceTimer,
 } from "../../utils/types";
 import type { ColoMetadata, Environment, ReadyAnalytics } from "./types";
@@ -13,7 +14,7 @@ import type { ColoMetadata, Environment, ReadyAnalytics } from "./types";
 interface Env {
 	ASSET_WORKER: Service<AssetWorker>;
 	USER_WORKER: Fetcher;
-	CONFIG: RoutingConfig;
+	CONFIG: RouterConfig;
 
 	SENTRY_DSN: string;
 	ENVIRONMENT: Environment;
@@ -52,6 +53,8 @@ export default {
 				env.CONFIG?.script_id
 			);
 
+			const config = applyConfigurationDefaults(env.CONFIG);
+
 			const url = new URL(request.url);
 
 			if (env.COLO_METADATA && env.VERSION_METADATA && env.CONFIG) {
@@ -66,7 +69,7 @@ export default {
 					coloRegion: env.COLO_METADATA.coloRegion,
 					hostname: url.hostname,
 					version: env.VERSION_METADATA.tag,
-					userWorkerAhead: env.CONFIG.invoke_user_worker_ahead_of_assets,
+					userWorkerAhead: config.invoke_user_worker_ahead_of_assets,
 				});
 			}
 
@@ -74,8 +77,8 @@ export default {
 
 			// User's configuration indicates they want user-Worker to run ahead of any
 			// assets. Do not provide any fallback logic.
-			if (env.CONFIG.invoke_user_worker_ahead_of_assets) {
-				if (!env.CONFIG.has_user_worker) {
+			if (config.invoke_user_worker_ahead_of_assets) {
+				if (!config.has_user_worker) {
 					throw new Error(
 						"Fetch for user worker without having a user worker binding"
 					);
@@ -96,12 +99,12 @@ export default {
 
 			// If we have a user-Worker, but no assets, dispatch to Worker script
 			const assetsExist = await env.ASSET_WORKER.unstable_canFetch(request);
-			if (env.CONFIG.has_user_worker && !assetsExist) {
+			if (config.has_user_worker && !assetsExist) {
 				analytics.setData({ dispatchtype: DISPATCH_TYPE.WORKER });
 
 				return await env.JAEGER.enterSpan("dispatch_worker", async (span) => {
 					span.setTags({
-						hasUserWorker: env.CONFIG.has_user_worker || false,
+						hasUserWorker: config.has_user_worker,
 						asset: assetsExist,
 						dispatchType: DISPATCH_TYPE.WORKER,
 					});
@@ -115,7 +118,7 @@ export default {
 			analytics.setData({ dispatchtype: DISPATCH_TYPE.ASSETS });
 			return await env.JAEGER.enterSpan("dispatch_assets", async (span) => {
 				span.setTags({
-					hasUserWorker: env.CONFIG.has_user_worker || false,
+					hasUserWorker: config.has_user_worker,
 					asset: assetsExist,
 					dispatchType: DISPATCH_TYPE.ASSETS,
 				});